+++ docs/CHANGE_LOG.md
@@ -4,6 +4,10 @@
 
 #### **Public Functions:**
 - Added "addCssToHead()" support for "html()", which will force all CSS style properties to be added as "style" tags to the head of the HTML document.
+- Added "indentSpaces()" support for "json()", which states the total number of indent spaces to use for the friendly format JSON returned from "get()".
+
+#### **Configuration Options:**
+- Removed the configuration option "jsonIndentationSpaces" (no longer needed).
 
 <br>